the directory life_lock has been renamed to smart_lock
