+++ locklib/__init__.py
@@ -1,3 +1,3 @@
-from locklib.locks.life_lock.lock import SmartLock
+from locklib.locks.smart_lock.lock import SmartLock
 
 from locklib.errors import DeadLockError
