+++ locklib/tests/life_lock/test_lock.py
@@ -4,7 +4,7 @@ from threading import Thread, Lock
 
 import pytest
 
-from locklib.locks.life_lock.lock import SmartLock
+from locklib.locks.smart_lock.lock import SmartLock
 from locklib.errors import DeadLockError
